Website-Analyzer-Agent/requirements.txt
requests 
beautifulsoup4 
builtwith 
python-whois 
tldextract